Зафиксил поведение инструментов в ресторане. Добавил статические крнструктуры, где необходимо. Разнес скрипты по папкам + неймспейсы
