+++ Assets/AssetPacks/Animations/Gathering/MiningLoopRW.anim
@@ -95602,7 +95602,7 @@ AnimationClip:
   m_HasMotionFloatCurves: 0
   m_Events:
   - time: 1.2666667
-    functionName: UnequipTool
+    functionName: Unequip
     data: 
     objectReferenceParameter: {fileID: 0}
     floatParameter: 0
